+++ src/features/LibraryModal/AssignKnowledgeBase/Item/index.tsx
@@ -57,7 +57,7 @@ const PluginItem = memo<KnowledgeItem>(({ id, fileType, name, type, description,
           )}
         </Flexbox>
       </Flexbox>
-      <Actions enabled={enabled} id={id} type={type} />
+      <Actions enabled={enabled} fileType={fileType} id={id} name={name} type={type} />
     </Flexbox>
   );
 });